Fix node registration to support multiple nodes with unique IP:port identification
## Problem
Two server nodes on different PCs were registering with the same address, causing
error 2067 (UNIQUE constraint violation) when the second node tried to register
with an empty wallet_address.

## Solution

### Database Schema (Central Server)
- Added ip_address and port columns to nodes table for endpoint tracking
- Removed inline UNIQUE constraint from wallet_address column
- Created partial UNIQUE index on wallet_address (only for non-empty values)
- Allows multiple test nodes with empty wallet addresses
- Created migrations:
  - 20250118000001_add_node_endpoint.sql
  - 20250118000002_fix_wallet_unique_constraint.sql

### Node Registration Logic (Central Server)
- Implemented smart duplicate detection:
  - Primary: Check by wallet_address (for authenticated nodes)
  - Fallback: Check by ip_address:port (for unauthenticated test nodes)
- Update existing node's IP:port when wallet address matches (handles IP changes)
- Store node endpoint information during registration
- Added new database queries:
  - get_node_by_endpoint()
  - update_node_endpoint()

### Code Quality (Backend)
- Fixed size_t conversion warnings in tensor.cpp and device.cpp
- Added explicit static_cast for ArrayFire dim4 operations
- Added explicit static_cast for memory calculations

### Engine Improvements
- Added connection dialog for Central Server
- Enhanced job management and monitoring
- Improved gRPC client connectivity

### Documentation & Scripts
- Created NODE_ENDPOINT_FIX.md with comprehensive fix documentation
- Added batch files for easy server management:
  - start-central-server.bat
  - start-server-node.bat
  - stop-all.bat

## Testing
Run two server nodes from different PCs:
- PC 1 (192.168.1.10) → Registers as unique node
- PC 2 (192.168.1.20) → Registers as different unique node
- Both nodes co-exist with empty wallet addresses
- No more error 2067 UNIQUE constraint failures

+++ cyxwiz-backend/src/core/device.cpp
@@ -122,14 +122,14 @@ DeviceInfo Device::GetInfo() const {
                 // Try to extract memory size from device name (e.g., "GTX 1050 Ti 4GB")
                 if (device_name.find("1050") != std::string::npos) {
                     info.memory_total = 4LL * 1024 * 1024 * 1024; // 4 GB
-                    info.memory_available = info.memory_total * 0.9; // Estimate 90% available
+                    info.memory_available = static_cast<size_t>(info.memory_total * 0.9); // Estimate 90% available
                     spdlog::info("Using estimated memory for {}: {} GB", device_name,
                         info.memory_total / (1024.0 * 1024.0 * 1024.0));
                 } else if (device_name.find("UHD") != std::string::npos ||
                            device_name.find("Intel") != std::string::npos) {
                     // Integrated Intel graphics typically share system RAM
                     info.memory_total = 2LL * 1024 * 1024 * 1024; // 2 GB shared
-                    info.memory_available = info.memory_total * 0.8;
+                    info.memory_available = static_cast<size_t>(info.memory_total * 0.8);
                     spdlog::info("Using estimated memory for {}: {} GB shared", device_name,
                         info.memory_total / (1024.0 * 1024.0 * 1024.0));
                 }
